update api cart

+++ bee-shoes-be/src/main/java/com/poly/tool/DBGenerator.java
@@ -206,7 +206,7 @@ public class DBGenerator implements CommandLineRunner {
         promotion1.setCode("PRO1");
         promotion1.setName("Khuyến mại % TEST");
         promotion1.setStatus(1);
-        promotion1.setValue(BigDecimal.valueOf(5));
+        promotion1.setValue(5);
         promotion1.setStartDate(LocalDateTime.of(2023, Month.OCTOBER, 15, 14, 0, 0));
         promotion1.setEndDate(LocalDateTime.of(2033, Month.OCTOBER, 15, 14, 0, 0));
         promotion1.setId(promotionRepository.save(promotion1).getId());
@@ -215,7 +215,7 @@ public class DBGenerator implements CommandLineRunner {
         promotion2.setCode("PRO2");
         promotion2.setName("Khuyến mại TEST");
         promotion2.setStatus(1);
-        promotion2.setValue(BigDecimal.valueOf(10));
+        promotion2.setValue(10);
         promotion2.setStartDate(LocalDateTime.of(2023, Month.OCTOBER, 15, 14, 0, 0));
         promotion2.setEndDate(LocalDateTime.of(2033, Month.OCTOBER, 15, 14, 0, 0));
         promotion2.setId(promotionRepository.save(promotion2).getId());
